requirements.txt
binance-connector
